More semantic path points for smiley
By making the eyes two points you can see at a glance which are the eyes and which is the mouth, it's a bit more intuitive. Also, you can see that the line component can parse out 2 or 3 or however many points. 

Produces the same result, would have to update the codepen too...